src/battles_in_antiquity/requirements.txt
black
pandas
pillow
pytest
git+https://github.com/jukujala/tgc-utils@main#egg=tgc_utils
git+https://github.com/jukujala/tts-utils@main#egg=tts_utils
